Menyelesaikan API
Progress:
 - Menyelesaikan Endpoint /setting/post
 - Menyelesaikan Endpoint /setting/get
 - Menyelesaikan Endpoint /sensor/post
 - Menyelesaikan Endpoint /sensor/get
 - Menambah seeder dengan fixture
 - Menambahkan Endpoint /sensor/get/sehari
 - Edit Readme

Notes:
 - Kalau bisa progress dilaporkan kpd client
 - Segera integrate ke aplikasi flutter agar masalah bisa segera diketahui

+++ service_api/models.py
@@ -13,4 +13,4 @@ class Pengukuran(models.Model):
 
 class Konfigurasi(models.Model):
     nama_konfigurasi = models.CharField(max_length=30)
-    nilai = models.IntegerField()
+    nilai = models.FloatField()
